build fixes

+++ ResourceManager.cpp
@@ -83,7 +83,7 @@ bool ResourceManager::loadGeometryFromObj(const path& path, std::vector<VertexAt
 		size_t offset = vertexData.size();
 		vertexData.resize(offset + shape.mesh.indices.size());
 
-		for (int i = 0; i < vertexData.size(); ++i) {
+		for (size_t i = 0; i < vertexData.size(); ++i) {
 			const tinyobj::index_t& idx = shape.mesh.indices[i];
 
 			vertexData[offset + i].position = {
